# website/public/openapi/serve.openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "EventsIngestRequest": {
        "properties": {
          "events": {
            "description": "Batch of interaction events",
            "items": {
              "$ref": "#/components/schemas/InteractionEvent"
            },
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "EventsIngestRequest",
        "type": "object"
      },
      "EventsIngestResponse": {
        "properties": {
          "accepted": {
            "description": "Number of events accepted into the source queue",
            "title": "Accepted",
            "type": "integer"
          },
          "event_ids": {
            "description": "Assigned or provided event ids",
            "items": {
              "type": "string"
            },
            "title": "Event Ids",
            "type": "array"
          }
        },
        "required": [
          "accepted",
          "event_ids"
        ],
        "title": "EventsIngestResponse",
        "type": "object"
      },
...
      "InteractionEvent": {
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Optional idempotency id",
            "title": "Event Id"
          },
          "event_type": {
            "description": "Must match a key in features.toml [event_weights] to affect weights",
            "title": "Event Type",
            "type": "string"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "description": "Alias for event_id",
            "title": "Idempotency Key"
          },
          "item_id": {
            "description": "Item identifier",
            "title": "Item Id",
            "type": "string"
          },
          "occurred_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              }
            ],
            "description": "ISO-8601 timestamp with timezone (Z or explicit offset), or Unix epoch seconds (UTC); converted to UTC",
            "examples": [
              "2026-08-13T12:00:00Z",
              1724000000
            ],
            "title": "Occurred At"
          },
          "quantity": {
            "default": 1,
            "description": "Optional count; default 1",
            "minimum": 1,
            "title": "Quantity",
            "type": "integer"
          },
          "user_id": {
            "description": "User identifier",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "item_id",
          "event_type",
          "occurred_at"
        ],
        "title": "InteractionEvent",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Read-only HTTP API over **precomputed** recommendations written by the batch job.\n\nThere is no live inference in the request path: `GET /recommendations/{user_id}`\nlooks up rows already stored in the configured output (dataset parquet or DB).\n\nWhen `[events]` is enabled with `kind = \"webhook\"`, `POST /events` accepts\ninteraction events for micro-batch incremental updates (write-through to the\nsame output store). See `docs/incremental-events.md`.\n\nWhen `[track]` is enabled, `POST /track` accepts recommendation impressions\nand clicks (not used for training). See `docs/evaluation.md`.\n\nInteractive docs: `/docs` (Swagger UI) and `/redoc` (includes language\ncode samples via ``x-codeSamples``). Machine-readable schema: `/openapi.json`.\nA checked-in copy lives at `docs/openapi/serve.openapi.json` (regenerate with\n`cicerone export-openapi`).",
    "title": "Cicerone Serve API",
    "version": "0.8.1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/events": {
      "post": {
        "operationId": "post_events_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/InteractionEvent"
                  },
                  {
                    "items": {
                      "$ref": "#/components/schemas/InteractionEvent"
                    },
                    "type": "array"
                  },
                  {
                    "$ref": "#/components/schemas/EventsIngestRequest"
                  }
                ],
                "title": "EventsIngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorDetail"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorDetail"
                    }
                  ],
                  "title": "Response 400 Post Events Events Post"
                }
              }
            },
            "description": "Invalid event payload: string ``ErrorDetail`` for malformed JSON / normalize errors, or ``ValidationErrorDetail`` (Pydantic ``errors()`` list)"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Request body too large"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Event backlog full"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest interaction events for incremental updates",
        "tags": [
          "events"
        ],
        "x-codeSamples": [
          {
            "label": "Net::HTTP",
            "lang": "Ruby",
            "source": "require \"json\"\nrequire \"net/http\"\nrequire \"uri\"\n\nbase = ENV.fetch(\"CICERONE_SERVE_URL\", \"http://localhost:8000\").sub(%r{/\\z}, \"\")\ntoken = ENV.fetch(\"CICERONE_SERVE_TOKEN\")\nuri = URI(\"#{base}/events\")\nbody = {\n  \"user_id\" => ENV.fetch(\"CICERONE_USER_ID\", \"alice\"),\n  \"item_id\" => \"ipa-001\",\n  \"event_type\" => \"purchase\",\n  \"quantity\" => 1,\n  \"occurred_at\" => \"2026-08-19T12:00:00Z\",\n}.to_json\nreq = Net::HTTP::Post.new(uri)\nreq[\"Authorization\"] = \"Bearer #{token}\"\nreq[\"Content-Type\"] = \"application/json\"\nreq[\"Accept\"] = \"application/json\"\nreq.body = body\nres = Net::HTTP.start(uri.hostname, uri.port, use_ssl: uri.scheme == \"https\") do |http|\n  http.request(req)\nend\nabort(\"HTTP #{res.code}: #{res.body}\") unless res.is_a?(Net::HTTPSuccess)\nputs JSON.parse(res.body)\n"
          },
          {
            "label": "urllib",
            "lang": "Python",
            "source": "import json\nimport os\nfrom urllib.request import Request, urlopen\n\nbase = os.environ.get(\"CICERONE_SERVE_URL\", \"http://localhost:8000\").rstrip(\"/\")\ntoken = os.environ[\"CICERONE_SERVE_TOKEN\"]\npayload = {\n    \"user_id\": os.environ.get(\"CICERONE_USER_ID\", \"alice\"),\n    \"item_id\": \"ipa-001\",\n    \"event_type\": \"purchase\",\n    \"quantity\": 1,\n    \"occurred_at\": \"2026-08-19T12:00:00Z\",\n}\nrequest = Request(\n    f\"{base}/events\",\n    data=json.dumps(payload).encode(),\n    headers={\n        \"Authorization\": f\"Bearer {token}\",\n        \"Content-Type\": \"application/json\",\n        \"Accept\": \"application/json\",\n    },\n    method=\"POST\",\n)\nprint(json.load(urlopen(request)))\n"
          },
          {
            "label": "fetch",
            "lang": "JavaScript",
            "source": "const baseUrl = (process.env.CICERONE_SERVE_URL || \"http://localhost:8000\").replace(/\\/$/, \"\");\nconst token = process.env.CICERONE_SERVE_TOKEN;\nif (!token) {\n  console.error(\"Set CICERONE_SERVE_TOKEN to a bearer token\");\n  process.exit(1);\n}\nconst userId = process.env.CICERONE_USER_ID || \"alice\";\n\nasync function main() {\n  const response = await fetch(`${baseUrl}/events`, {\n    method: \"POST\",\n    headers: {\n      Accept: \"application/json\",\n      Authorization: `Bearer ${token}`,\n      \"Content-Type\": \"application/json\",\n    },\n    body: JSON.stringify({\n      user_id: userId,\n      item_id: \"ipa-001\",\n      event_type: \"purchase\",\n      quantity: 1,\n      occurred_at: \"2026-08-19T12:00:00Z\",\n    }),\n  });\n  if (!response.ok) throw new Error(`${response.status} ${await response.text()}`);\n  console.log(await response.json());\n}\n\nmain().catch((err) => {\n  console.error(err);\n  process.exitCode = 1;\n});\n"
          },
          {
            "label": "curl",
            "lang": "Shell",
            "source": "BASE_URL=\"${CICERONE_SERVE_URL:-http://localhost:8000}\"\nBASE_URL=\"${BASE_URL%/}\"\nTOKEN=\"${CICERONE_SERVE_TOKEN:?set CICERONE_SERVE_TOKEN}\"\nUSER_ID=\"${CICERONE_USER_ID:-alice}\"\nif command -v python3 >/dev/null 2>&1; then PYTHON=python3\nelif command -v python >/dev/null 2>&1; then PYTHON=python\nelse PYTHON=\nfi\nif [ -n \"$PYTHON\" ]; then\n  BODY=\"$(\n    USER_ID=\"$USER_ID\" \"$PYTHON\" -c \\\n      'import json, os\nprint(json.dumps({\n    \"user_id\": os.environ[\"USER_ID\"],\n    \"item_id\": \"ipa-001\",\n    \"event_type\": \"purchase\",\n    \"quantity\": 1,\n    \"occurred_at\": \"2026-08-19T12:00:00Z\",\n}))'\n  )\"\nelif command -v jq >/dev/null 2>&1; then\n  BODY=\"$(\n    jq -n --arg user_id \"$USER_ID\" \\\n      '{user_id:$user_id,item_id:\"ipa-001\",event_type:\"purchase\",quantity:1,occurred_at:\"2026-08-19T12:00:00Z\"}'\n  )\"\nelse\n  echo \"python3, python, or jq required to JSON-encode USER_ID\" >&2\n  exit 1\nfi\ncurl -fsS -X POST \\\n  -H \"Authorization: Bearer $TOKEN\" \\\n  -H \"Content-Type: application/json\" \\\n  -d \"$BODY\" \\\n  \"$BASE_URL/events\" || exit 1\n"
          }
        ]
      }
    },
...
    "/track": {
      "post": {
        "operationId": "post_track_track_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/TrackEvent"
                  },
                  {
                    "items": {
                      "$ref": "#/components/schemas/TrackEvent"
                    },
                    "type": "array"
                  },
                  {
                    "$ref": "#/components/schemas/TrackIngestRequest"
                  }
                ],
                "title": "TrackIngestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackIngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorDetail"
                    },
                    {
                      "$ref": "#/components/schemas/ValidationErrorDetail"
                    }
                  ],
                  "title": "Response 400 Post Track Track Post"
                }
              }
            },
            "description": "Invalid track payload"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Missing or invalid bearer token"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Request body too large"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ingest recommendation impressions and clicks",
        "tags": [
          "track"
        ],
        "x-codeSamples": [
          {
            "label": "Net::HTTP",
            "lang": "Ruby",
            "source": "require \"json\"\nrequire \"net/http\"\nrequire \"uri\"\n\nbase = ENV.fetch(\"CICERONE_SERVE_URL\", \"http://localhost:8000\").sub(%r{/\\z}, \"\")\ntoken = ENV.fetch(\"CICERONE_SERVE_TOKEN\")\nuri = URI(\"#{base}/track\")\nbody = {\n  \"user_id\" => ENV.fetch(\"CICERONE_USER_ID\", \"alice\"),\n  \"item_id\" => \"ipa-001\",\n  \"kind\" => \"impression\",\n  \"rank\" => 1,\n  \"occurred_at\" => \"2026-08-19T12:00:00Z\",\n}.to_json\nreq = Net::HTTP::Post.new(uri)\nreq[\"Authorization\"] = \"Bearer #{token}\"\nreq[\"Content-Type\"] = \"application/json\"\nreq[\"Accept\"] = \"application/json\"\nreq.body = body\nres = Net::HTTP.start(uri.hostname, uri.port, use_ssl: uri.scheme == \"https\") do |http|\n  http.request(req)\nend\nabort(\"HTTP #{res.code}: #{res.body}\") unless res.is_a?(Net::HTTPSuccess)\nputs JSON.parse(res.body)\n"
          },
          {
            "label": "urllib",
            "lang": "Python",
            "source": "import json\nimport os\nfrom urllib.request import Request, urlopen\n\nbase = os.environ.get(\"CICERONE_SERVE_URL\", \"http://localhost:8000\").rstrip(\"/\")\ntoken = os.environ[\"CICERONE_SERVE_TOKEN\"]\npayload = {\n    \"user_id\": os.environ.get(\"CICERONE_USER_ID\", \"alice\"),\n    \"item_id\": \"ipa-001\",\n    \"kind\": \"impression\",\n    \"rank\": 1,\n    \"occurred_at\": \"2026-08-19T12:00:00Z\",\n}\nrequest = Request(\n    f\"{base}/track\",\n    data=json.dumps(payload).encode(),\n    headers={\n        \"Authorization\": f\"Bearer {token}\",\n        \"Content-Type\": \"application/json\",\n        \"Accept\": \"application/json\",\n    },\n    method=\"POST\",\n)\nprint(json.load(urlopen(request)))\n"
          },
          {
            "label": "fetch",
            "lang": "JavaScript",
            "source": "const baseUrl = (process.env.CICERONE_SERVE_URL || \"http://localhost:8000\").replace(/\\/$/, \"\");\nconst token = process.env.CICERONE_SERVE_TOKEN;\nif (!token) {\n  console.error(\"Set CICERONE_SERVE_TOKEN to a bearer token\");\n  process.exit(1);\n}\nconst userId = process.env.CICERONE_USER_ID || \"alice\";\n\nasync function main() {\n  const response = await fetch(`${baseUrl}/track`, {\n    method: \"POST\",\n    headers: {\n      Accept: \"application/json\",\n      Authorization: `Bearer ${token}`,\n      \"Content-Type\": \"application/json\",\n    },\n    body: JSON.stringify({\n      user_id: userId,\n      item_id: \"ipa-001\",\n      kind: \"impression\",\n      rank: 1,\n      occurred_at: \"2026-08-19T12:00:00Z\",\n    }),\n  });\n  if (!response.ok) throw new Error(`${response.status} ${await response.text()}`);\n  console.log(await response.json());\n}\n\nmain().catch((err) => {\n  console.error(err);\n  process.exitCode = 1;\n});\n"
          },
          {
            "label": "curl",
            "lang": "Shell",
            "source": "BASE_URL=\"${CICERONE_SERVE_URL:-http://localhost:8000}\"\nBASE_URL=\"${BASE_URL%/}\"\nTOKEN=\"${CICERONE_SERVE_TOKEN:?set CICERONE_SERVE_TOKEN}\"\nUSER_ID=\"${CICERONE_USER_ID:-alice}\"\nif command -v python3 >/dev/null 2>&1; then PYTHON=python3\nelif command -v python >/dev/null 2>&1; then PYTHON=python\nelse PYTHON=\nfi\nif [ -n \"$PYTHON\" ]; then\n  BODY=\"$(\n    USER_ID=\"$USER_ID\" \"$PYTHON\" -c \\\n      'import json, os\nprint(json.dumps({\n    \"user_id\": os.environ[\"USER_ID\"],\n    \"item_id\": \"ipa-001\",\n    \"kind\": \"impression\", \"rank\": 1,\n    \"occurred_at\": \"2026-08-19T12:00:00Z\",\n}))'\n  )\"\nelif command -v jq >/dev/null 2>&1; then\n  BODY=\"$(\n    jq -n --arg user_id \"$USER_ID\" \\\n      '{user_id:$user_id,item_id:\"ipa-001\",kind:\"impression\",rank:1,occurred_at:\"2026-08-19T12:00:00Z\"}'\n  )\"\nelse\n  echo \"python3, python, or jq required to JSON-encode USER_ID\" >&2\n  exit 1\nfi\ncurl -fsS -X POST \\\n  -H \"Authorization: Bearer $TOKEN\" \\\n  -H \"Content-Type: application/json\" \\\n  -d \"$BODY\" \\\n  \"$BASE_URL/track\" || exit 1\n"
          }
        ]
      }
    }
  }
}